[level unrated] Title: 배열의 평균값, Time: 0.02 ms, Memory: 75.9 MB -BaekjoonHub

+++ 프로그래머스/unrated/120817. 배열의 평균값/배열의 평균값.java
@@ -1,11 +1,11 @@
 class Solution {
     public double solution(int[] numbers) {
         double answer = 0;
-        int sum = 0;
+        double sum = 0;
         for(int item : numbers){
             sum += item;
         }
-        answer = (double)sum / numbers.length;
+        answer = sum / numbers.length;
         return answer;
     }
 }
